requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
langchain==0.0.340
langchain-community==0.0.1
langchain-openai==0.0.2
qdrant-client==1.7.3
mistralai==0.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
openai==1.3.7
pydantic==2.4.2
Pillow==10.1.0